+++ README.md
@@ -1,5 +1,9 @@
 # PDFMergeXpress
 
-![App Screenshot](img/screenshot1-development.png)
+![App Screenshot](assets/img/docs/screenshot1-development.png)
 
 A simple PDF merging app made with customtkinter library.
+
+## Contributing
+
+Still a learning developer. You may read [here](/docs/_index.md) and see how it goes for now.
