+++ apps/components/converter_head.py
@@ -1,0 +1,27 @@
+# A combination of a progress bar, button, and my very own searchbar widget.
+from typing import Optional, Tuple, Union
+from customtkinter import *
+from .searchbar import Searchbar
+from .converter import Converter
+
+
+class ConverterHead(CTkFrame):
+    WIDTH: int = 500
+    HEIGHT: int = 50
+
+    def __init__(self, *args,
+                 bg_color: str | Tuple[str, str] = 'transparent',
+                 fg_color: str | Tuple[str, str] = 'whitesmoke',
+                 corner_radius: int = 0,
+                 **kwargs):
+        super().__init__(*args, bg_color=bg_color, fg_color=fg_color,
+                         corner_radius=corner_radius, **kwargs)
+
+        self.grid_columnconfigure(0, weight=1)
+
+        self.searchbar = Searchbar(master=self)
+        self.searchbar.grid(row=0, column=0, padx=10, pady=(5, 0))
+
+        self.converter = Converter(
+            master=self, get_path_func=self.searchbar.get)
+        self.converter.grid(row=1, column=0, padx=10, pady=(0, 5))
