+++ apps/components/merger_head.py
@@ -2,10 +2,10 @@
 from typing import Optional, Tuple, Union
 from customtkinter import *
 from .searchbar import Searchbar
-from .converter import Converter
+from .merger import Merger
 
 
-class HeadComponent(CTkFrame):
+class MergerHead(CTkFrame):
     WIDTH: int = 500
     HEIGHT: int = 50
 
@@ -22,20 +22,6 @@ class HeadComponent(CTkFrame):
         self.searchbar = Searchbar(master=self)
         self.searchbar.grid(row=0, column=0, padx=10, pady=(5, 0))
 
-        self.converter = Converter(
-            master=self, get_path_func=self.searchbar.get)
-        self.converter.grid(row=1, column=0, padx=10, pady=(0, 5))
-
-
-if __name__ == '__main__':
-    class Tester(CTk):
-        def __init__(self, fg_color: str | Tuple[str, str] | None = None, **kwargs):
-            super().__init__(fg_color, **kwargs)
-
-            self.grid_columnconfigure(0, weight=1)
-
-            self.head = HeadComponent(self)
-            self.head.grid(row=0, column=0)
-
-    app = Tester()
-    app.mainloop()
+        self.merger = Merger(master=self, get_path_func=self.searchbar.get,
+                             clear_search_func=self.searchbar.clear)
+        self.merger.grid(row=1, column=0, padx=10, pady=(0, 5))
