+++ apps/components/searchbar.py
@@ -52,22 +52,3 @@ class Searchbar(CTkFrame):
     def clear(self):
         self.entry.delete(0, 'end')
         self.entry.configure(placeholder_text='Folder Path')
-
-
-if __name__ == '__main__':
-    class Tester(CTk):
-        def __init__(self, fg_color: str | Tuple[str, str] | None = None, **kwargs):
-            super().__init__(fg_color, **kwargs)
-
-            self.grid_columnconfigure(0, weight=1)
-
-            self.head_frame = CTkFrame(
-                self, fg_color='whitesmoke', corner_radius=0)
-            self.head_frame.grid(row=0, column=0)
-
-            self.searchbar = Searchbar(master=self.head_frame, width=700)
-            self.searchbar.grid(row=0, column=0, padx=10, pady=(10, 0))
-
-    app = Tester()
-
-    app.mainloop()
