--- components/__init__.py
@@ -1,5 +1,0 @@
-from .merger_head import HeadComponent
-from .treeview import TreeView
-from .appearance_mode_toggler import AppearanceModeToggler
-from .searchbar import Searchbar
-from .merger import Merger
