+++ utils/__init__.py
@@ -1,1 +1,8 @@
-from .logics import LogicHandler
+from .pdf_merger import PdfMerger
+from .docx_to_pdf import DocxConverter
+
+
+__all__ = [
+    'PdfMerger',
+    'DocxConverter',
+]
