+++ utils/pdf_merger.py
@@ -3,10 +3,11 @@ from PyPDF2 import PdfMerger, PdfFileReader
 import os
 import glob
 
-class LogicHandler:
+
+class PdfMerger:
     def __init__(self) -> None:
         pass
-    
+
     @staticmethod
     def get_pdf_page(file_path: str) -> int:
         """Returns the number of pages of a given PDF file
@@ -18,40 +19,43 @@ class LogicHandler:
             int: total number of pages
         """
         file = open(file_path, 'rb')
-        
-        # store data in pdfReader 
-        pdf_reader = PdfFileReader(file) 
-        
-        # count number of pages 
-        total_pages = len(pdf_reader.pages) 
-        
-        return  total_pages
+
+        # store data in pdfReader
+        pdf_reader = PdfFileReader(file)
+
+        # count number of pages
+        total_pages = len(pdf_reader.pages)
+
+        return total_pages
 
     @staticmethod
     def get_pdfs(root_dir: list) -> list:
         files = []
         for row_num, folder in enumerate(root_dir):
-            str_path = os.path.join(folder, '*.pdf') # Use os.path.join for paths
-            pdfs = [path.replace("\\","/") for path in glob.glob(str_path, recursive=True) if os.path.isfile(path)] # also check if the glob is a pdf file, it may sometimes be a folder that ends with .pdf
+            # Use os.path.join for paths
+            str_path = os.path.join(folder, '*.pdf')
+            # also check if the glob is a pdf file, it may sometimes be a folder that ends with .pdf
+            pdfs = [path.replace(
+                "\\", "/") for path in glob.glob(str_path, recursive=True) if os.path.isfile(path)]
             # .replace("\\","/") for 'slash' consistency... https://stackoverflow.com/a/18776536, tnx
             files.append(pdfs)
-            
+
     @staticmethod
     #  Credits to Ma'am Tine (https://github.com/kjvmartinez) for this function
     def merge_pdf(pdfs: list[str], filename: str):
-            if len(pdfs) == 1:  # If there's only one PDF file, skip merging
-                return 0
+        if len(pdfs) == 1:  # If there's only one PDF file, skip merging
+            return 0
 
-            merger = PdfMerger()
-            for pdf_file in pdfs:
-                merger.append(pdf_file)
-                
-                # merger.append(i)
-            merger.write(filename)
-            num_pages = len(merger.pages)
-            merger.close()
-            return num_pages
-        
+        merger = PdfMerger()
+        for pdf_file in pdfs:
+            merger.append(pdf_file)
+
+            # merger.append(i)
+        merger.write(filename)
+        num_pages = len(merger.pages)
+        merger.close()
+        return num_pages
+
     @staticmethod
     # Credits to jbarry302 (https://github.com/jbarry302) for all the functions below
     def get_num_pages(pdfs: list) -> int:
@@ -63,38 +67,38 @@ class LogicHandler:
             int: The total number of pages in all the PDF files combined.
         """
         merger = PdfMerger()
-        
+
         for pdf_file in pdfs:
             merger.append(pdf_file)
 
         num_pages = len(merger.pages)
-        
+
         merger.close()
-        
+
         return num_pages
 
     def flatten_array(self, arr: list) -> list:
         """
         Recursively flattens a nested list into a single list.
-        
+
         Args:
             arr (list): A nested list to be flattened.
-        
+
         Returns:
             list: A flattened list.
-        
+
         Example:
         >>> flatten_array([1, [2, [3, 4], 5], 6])
         [1, 2, 3, 4, 5, 6]
         """
         flattened_list = []
-        
+
         for element in arr:
             if isinstance(element, list):
                 flattened_list.extend(self.flatten_array(element))
             else:
                 flattened_list.append(element)
-                
+
         return flattened_list
 
     @staticmethod
@@ -119,8 +123,11 @@ class LogicHandler:
         files = []
 
         for row_num, folder in enumerate(root_dir):
-            str_path = os.path.join(folder, '*.pdf') # Use os.path.join for paths
-            pdfs = [path.replace("\\","/") for path in glob.glob(str_path, recursive=True) if os.path.isfile(path)] # also check if the glob is a pdf file, it may sometimes be a folder that ends with .pdf
+            # Use os.path.join for paths
+            str_path = os.path.join(folder, '*.pdf')
+            # also check if the glob is a pdf file, it may sometimes be a folder that ends with .pdf
+            pdfs = [path.replace(
+                "\\", "/") for path in glob.glob(str_path, recursive=True) if os.path.isfile(path)]
             # .replace("\\","/") for 'slash' consistency... https://stackoverflow.com/a/18776536, tnx
             files = files + pdfs
             number_of_pages = self.get_num_pages(pdfs)
@@ -130,7 +137,7 @@ class LogicHandler:
         # by default, save the output to the first folder
         output_path = os.path.join(root_dir[0], 'output.pdf')
         self.merge_pdf(files, output_path)
-        
+
     @staticmethod
     def scan_folders(folder_path: str) -> list:
         list_roots = []
@@ -138,24 +145,28 @@ class LogicHandler:
         for (root, dirs, files) in os.walk(folder_path, topdown=True):
             if any(f.endswith('.pdf') or f.endswith('.PDF') for f in files):
                 list_roots.append(root)
-        
+
         return list_roots
 
     # This ugly function is made by me
     def deep_merging(self, main_path: str, dirs: list) -> tuple[list]:
         outputs = []
-        
+
         for folder in dirs:
-            str_path = os.path.join(folder, '*.pdf') # Use os.path.join for paths
-            pdfs = [path for path in glob.glob(str_path, recursive=True) if os.path.isfile(path) and not os.path.basename(path).startswith('output')]       
-            
-            number_of_pages = self.merge_pdf(pdfs, os.path.join(folder, 'output.pdf'))
-        
-            if number_of_pages == 0: continue
-            
+            # Use os.path.join for paths
+            str_path = os.path.join(folder, '*.pdf')
+            pdfs = [path for path in glob.glob(str_path, recursive=True) if os.path.isfile(
+                path) and not os.path.basename(path).startswith('output')]
+
+            number_of_pages = self.merge_pdf(
+                pdfs, os.path.join(folder, 'output.pdf'))
+
+            if number_of_pages == 0:
+                continue
+
             folder_path = f"📁 {os.path.normpath(folder)}"
             link = os.path.join(folder, 'output.pdf')
-            
+
             outputs.append((link, folder_path, number_of_pages))
-            
+
         return outputs
